Allow explicit declaration of people/super properties
This change does not affect default behavior; as long as a user has "Automatically set all Traits as Super Properties and People Properties" checked in our UI (which it is by default), the integration will behave as it always has, setting all traits to super properties and all traits to People Properties if "Use Mixpanel People" is enabled as well.

However, this change allows our users to *also* pass optional options.Mixpanel.peopleProperties and options.Mixpanel.superProperties arrays in both `track` and `identify` that delineate which `.properties` or `.traits` fields respectively that they'd like to pass to which mixpanel API. This enables two new use cases in particular:

- Setting super properties on `track` calls (granularly to boot).
- Overriding default trait handling (sending all as both) and instead specifying which traits to set as super properties and which to set as people properties specifically.

The commit history here is a bit muddled as we balked on a prior implementation with similar goals that was erroneously merged to master and released (1.1.0).